+++ Sources/Models/ChainHead.swift
@@ -1,0 +1,15 @@
+//
+//  ChainHead.swift
+//  TezosSwift
+//
+//  Created by Marek Fořt on 11/5/18.
+//  Copyright © 2018 Keefer Taylor. All rights reserved.
+//
+
+import Foundation
+
+public struct ChainHead: Codable {
+    let chainId: String
+    let hash: String
+    let `protocol`: String
+}
