+++ Sources/Models/ContractStatus.swift
@@ -26,20 +26,3 @@ extension ContractStatus: Decodable {
         self.init(balance: balance, spendable: spendable)
     }
 }
-
-extension KeyedDecodingContainer {
-    func decodeRPC(_ type: Int.Type, forKey key: K) throws -> Int {
-        let intString = try decode(String.self, forKey: key)
-        let context = DecodingError.Context(codingPath: codingPath, debugDescription: "Decryption failed")
-        guard let decodedInt = Int(intString) else { throw DecodingError.dataCorrupted(context) }
-        return decodedInt
-    }
-}
-
-
-public struct ChainHead: Codable {
-    let chainId: String
-    let hash: String
-    let `protocol`: String
-}
-
