+++ Sources/Models/SignedOperationPayload.swift
@@ -1,0 +1,16 @@
+//
+//  SignedOperationPayload.swift
+//  TezosSwift
+//
+//  Created by Marek Fořt on 11/5/18.
+//  Copyright © 2018 Keefer Taylor. All rights reserved.
+//
+
+import Foundation
+
+struct SignedOperationPayload: Encodable {
+    let contents: [Operation]
+    let branch: String
+    let `protocol`: String
+    let signature: String
+}
